Update battle sequence, refactor helpers, begin adding front end

+++ java/com/game/troops/Troops.java
@@ -6,6 +6,6 @@ public interface Troops {
 	public int getArmor();
 	public int getSpeed();
 	public int getLoot();
-	public int getHealth();
+	public double getHealth();
 	public int getBattlePosition();
 }
